+++ src/view/pages/console/appstore/detail/detail.js
@@ -56,7 +56,8 @@ export default {
 
   created() {
     this.getApp();
-    this.getChart();
+    // this.getChart();
+    this.getCharts();
     this.getCategory();
     this.getInstances();
   },
@@ -79,20 +80,25 @@ export default {
       AppStoreService.getApp(this.zone.id, this.space.id, this.$route.params.Id).then(res => {
         if (res) {
           this.appInfo = res;
-
           this.form.category = res.categoryId;
-          this.form.name = `${res.name.split('-')[1]}`;
+
+          const length = `${res.name.split('-').length}`;
+          if (res.isGlobal || length < 2) {
+            this.form.name = res.name;
+          } else {
+            this.form.name = `${res.name.split('-')[1]}`;
+            // if (res.name.split('-'))
+          }
           this.form.description = res.description;
         }
       });
     },
-    // 获取chart
-    getChart() {
-      AppStoreService.getApp(this.zone.id, this.space.id, this.$route.params.Id).then(res => {
+    getCharts() {
+      AppStoreService.getCharts(this.zone.id, this.space.id, this.$route.params.Id).then(res => {
         if (res) {
-          this.applicationInfos = res.applicationInfos;
+          this.applicationInfos = res;
         }
-        res.applicationInfos.forEach(item => {
+        res.forEach(item => {
           this.chart = item.version;
         });
       });
@@ -122,6 +128,7 @@ export default {
             this.$noty.success('修改成功');
             this.editClose();
             this.getApp();
+            this.isShow = true;
           }
         });
     },
@@ -148,12 +155,15 @@ export default {
       this.isShow = !this.isShow;
     },
     // 更新表单
-    linktoForm() {
+    linktoForm(id) {
       this.$router.push({
         name: 'appstore.form',
         params: {
           appid: this.appInfo.id,
           version: this.chart,
+        },
+        query: {
+          instanceId: id,
         },
       });
     },
@@ -262,9 +272,6 @@ export default {
       this.configAdd = false;
     },
 
-    // removeTag(res) {
-    //   this.form.category = res;
-    // },
     // 获取实例数
     instanceNum() {
       return this.instanceTable.length;
@@ -290,30 +297,23 @@ export default {
     // 上传图片文件
     handleUpload() {
       this.isDisabled = true;
-      // const formData = new FormData();
-      // this.fileList.forEach(file => {
-      //   formData.append('blob', file);
-      // });
-      // AppStoreService.uploadImg(formData)
-      //   .then(res => {
-      //     this.form.pictureId = res.id;
-      //     this.updateApp();
-      //   })
-      //   .catch(err => {
-      //     this.removeFile();
-      //     this.$message.error(err);
-      //   });
-      const file = this.fileList[0];
-      AppStoreService.uploadPic(file)
-        .then(res => {
-          this.form.pictureId = res.id;
-          this.updateApp();
-        })
-        .catch(err => {
-          this.removeFile();
-          this.$message.error(err);
-        });
+      if (this.fileList.length) {
+        const file = this.fileList[0];
+        AppStoreService.uploadPic(file)
+          .then(res => {
+            this.form.pictureId = res.id;
+            this.updateApp();
+          })
+          .catch(err => {
+            this.removeFile();
+            this.$noty.error(err);
+          });
+      } else {
+        this.form.pictureId = this.appInfo.pictureId;
+        this.updateApp();
+      }
     },
+
     // 上传chart文件之前
     beforeUploadChart(file) {
       if (this.chartType.indexOf(file.type) < 0) {
@@ -336,6 +336,9 @@ export default {
         .then(res => {
           if (res) {
             this.$noty.success('上传chart成功');
+            this.getApp();
+            this.getCharts();
+            this.addClose();
           }
         })
         .catch(err => {
